Move cacheSize validation tests into reusable function
