+++ src/cache/__tests__/LruCacheObject.spec.js
@@ -1,11 +1,12 @@
 import CacheObject from '../LruCacheObject';
 import testBasicBehavior from '../__util__/testBasicBehavior';
 import testLruBehavior from '../__util__/testLruBehavior';
-import * as validateCacheSize from '../util/validateCacheSize';
+import testCacheSizeValidation from '../__util__/testCacheSizeValidation';
 
 describe('LruCacheObject', () => {
   testBasicBehavior(CacheObject, {cacheSize: 10});
   testLruBehavior(CacheObject);
+  testCacheSizeValidation(CacheObject);
 
   describe('isValidCacheKey', () => {
     it('Should accept only numbers and string', () => {
