+++ src/cache/__util__/testLruBehavior.js
@@ -1,5 +1,4 @@
 import fillCacheWith from './fillCacheWith';
-import * as validateCacheSize from '../util/validateCacheSize';
 
 function testLruBehavior(CacheObject) {
   describe('LRU cache behavior', () => {
@@ -32,15 +31,6 @@ function testLruBehavior(CacheObject) {
       newEntries.forEach(entry => {
         expect(cache.get(entry)).toBe(entry);
       });
-    });
-
-    // @TODO Test the feature, not implementation
-    it('Should validate `cacheSize` parameter', () => {
-      const spy = jest.spyOn(validateCacheSize, 'default');
-      const cache = new CacheObject({cacheSize: 5});
-
-      expect(spy).toHaveBeenCalledWith(5);
-      spy.mockRestore();
     });
   });
 }
